requirements.txt
pypng==0.0.18
pillow==4.1.1
